pixiv-tags/pyproject.toml
[project]
name = "pixiv-tags"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "python-dotenv>=1.2.1",
]
